% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{npz_file}
\alias{npz_file}
\title{A function to extract the om matrix from an npz calibration}
\usage{
npz_file(calibration_file)
}
\description{
A function to extract the om matrix from an npz calibration
}
