% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rodrigues}
\alias{rodrigues}
\title{A function to do the Rodrigues transform rotation on a matlab calibration}
\usage{
rodrigues(om)
}
\description{
Translated from Matlab Toolbox (Bouget et al. 2008). Requires a *.mat calibration file.
}
