% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SEB_data_concatenate_project}
\alias{SEB_data_concatenate_project}
\title{A function to concatenate SEBASTES data to .csv files}
\usage{
SEB_data_concatenate_project(project.dir)
}
\arguments{
\item{project.dir}{directory containing the multiple deployment files}
}
\description{
This function inputs a directory folder containing multiple stereo camera deployments,
reads all the *.db files and extracts the frame data and the target data from all the
deployments contained in the directory. It then concatenates these together, assigning
a unique identifier (deployment_ID) and outputs two .csv files, one containing target
data and one containing the frame data. Any accessory data from sensors will also be
included with the frame data. This function works with the new "project" style data.
}
\examples{
SEB_data_concatenate_project("C:/Users/rooperc/Desktop/Rockfish Projects/Longline Survey Gear Comparisons/TrigCamData/Gadus")
}
\keyword{SEBASTES,}
\keyword{camera,}
\keyword{concatenation}
\keyword{data}
\keyword{stereo}
