% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SEB_data_replace}
\alias{SEB_data_replace}
\title{A function to replace SEBASTES deployment data with updated versions}
\usage{
SEB_data_replace(new.data.dir, project.dir, name_start = "D")
}
\arguments{
\item{new.data.dir}{a directory containing the new data (*.db) files}

\item{project.dir}{directory containing original deployment files (images, logs, settings)}

\item{name_start}{a character defining the start of the deployment designation. In SEBASTES this is
typically the default "D" (e.g. D20210903-T223225)}
}
\description{
This inputs a folder containing *.db files from SEBASTES annotations and uses their
file name to overwrite the existing *.db files in deployment folders. This is a handy
function if you have multiple people working on SEBASTES deployments with multiple
copies of the images. Use this function to update the old data with the new data after
correcting identifications or remeasuring individuals or processing for different groups
of species.
}
\examples{
SEB_data_replace("C:/Users/rooperc/Desktop/Rockfish Projects/Longline Survey Gear Comparisons/TrigCamData/Gadus data/Gadus data","C:/Users/rooperc/Desktop/Rockfish Projects/Longline Survey Gear Comparisons/TrigCamData/Gadus")
}
\keyword{SEBASTES,}
\keyword{camera,}
\keyword{concatenation,}
\keyword{data}
\keyword{stereo}
\keyword{wrangling}
