% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SEB_get_sensor}
\alias{SEB_get_sensor}
\title{A function to concatenate sensor data from a project to attach to image data}
\usage{
SEB_get_sensor(project.dir, make_csv = FALSE)
}
\arguments{
\item{project.dir}{directory containing the multiple deployment files}

\item{make_csv}{do you want an output sensor.csv file (default = FALSE)}
}
\description{
This function inputs a directory folder containing multiple stereo camera deployments,
reads all the sensor data from the deployments contained in the directory.
It then concatenates these together, assigning #' a unique identifier (deployment_ID)
and can outputs a .csv files.
}
\examples{
SEB_get_sensor("C:/Users/rooperc/Desktop/Rockfish Projects/Longline Survey Gear Comparisons/TrigCamData/Gadus")
}
\keyword{SEBASTES,}
\keyword{camera,}
\keyword{concatenation}
\keyword{data}
\keyword{stereo}
