% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SEB_GPS_point}
\alias{SEB_GPS_point}
\title{A function to append GPS data to TrigCam SEBASTES deployment data}
\usage{
SEB_GPS_point(
  frame_data,
  Deployment_ID,
  Longitude,
  Latitude,
  Accessory_data = NULL
)
}
\arguments{
\item{frame_data}{a standard output from the SEB_data_concatenate function used on images
analyzed using the SEBASTES software.}

\item{Deployment_ID}{a column of deployment IDs corresponding to the unique IDs in the frame data}

\item{Longitude}{a column of longitudes for each deployment ID}

\item{Latitude}{a column of latitudes for each deployment ID}

\item{Accessory_data}{columns of accessory data to be attached to the frame data}
}
\description{
This function merges a GPS position to frame data from SEBASTES output for point data. The GPS data is a
single pair of Latitude and Longitude for each deployment (e.g. TrigCam deployments). A
third column identifies the Deployment_ID. Accessory data, such as depth inlcuded in the
data frame can also be appended to the output.
}
\examples{
SEB_data_replace(frame_data, GPS_data$Deployment_ID, GPS_data$Longitude, GPS_data$Latitude, GPS_data[,4:10])
}
\keyword{SEBASTES,}
\keyword{camera,}
\keyword{concatenation,}
\keyword{data}
\keyword{stereo}
\keyword{wrangling}
