% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SEB_GPS_transect}
\alias{SEB_GPS_transect}
\title{A function to append GPS data to transect SEBASTES deployment data}
\usage{
SEB_GPS_transect(
  frame_data,
  Deployment_ID,
  Time,
  Longitude,
  Latitude,
  Accessory_data = NULL,
  offset = 0
)
}
\arguments{
\item{frame_data}{a standard output from the SEB_data_concatenate function used on images
analyzed using the SEBASTES software.}

\item{Deployment_ID}{a column of deployment IDs corresponding to the unique IDs in the frame data}

\item{Time}{a column of times in as.POSIXct format (e.g. '\%Y-\%m-\%d \%H:\%M:\%OS')}

\item{Longitude}{a column of longitudes for each deployment ID}

\item{Latitude}{a column of latitudes for each deployment ID}

\item{Accessory_data}{columns of accessory data to be attached to the frame data}

\item{offset}{offset in hours for conversion between GPS timestamp and camera timestamp}
}
\description{
This function merges a GPS position to frame data from SEBASTES output for transect data. The GPS data is a
set of Latitude and Longitude pairs with an associated time for each deployment (e.g. a trackline for a drop
camera deployments). A third column identifies the Deployment_ID. Accessory data, such as depth included in the
data frame can also be appended to the output.
}
\examples{
SEB_data_replace(frame_data, GPS_data$Deployment_ID, GPS_data$Time, GPS_data$Longitude, GPS_data$Latitude, GPS_data[,4:10],offset=-7)
}
\keyword{SEBASTES,}
\keyword{camera,}
\keyword{concatenation,}
\keyword{data}
\keyword{stereo}
\keyword{wrangling}
