% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SEB_SeaBird_append}
\alias{SEB_SeaBird_append}
\title{A function to append Seabird SBE data to SEBASTES deployment data}
\usage{
SEB_SeaBird_append(frame_data, SBE_file, offset = 0)
}
\arguments{
\item{frame_data}{a standard output from the SEB_data_concatenate function used on images
analyzed using the SEBASTES software.}

\item{SBE_file}{a single or multiple SBE files in their native format}

\item{offset}{offset in hours for conversion between SBE timestamp and camera timestamp}
}
\description{
This function merges a Seabird SBE file with time, depth and temperature to frame data from SEBASTES output.
}
\examples{
SEB_SeaBird_append(frame_data, SBE_file,offset=-7)
}
\keyword{SEBASTES,}
\keyword{SeaBird}
\keyword{camera,}
\keyword{concatenation,}
\keyword{data}
\keyword{stereo}
\keyword{wrangling,}
