% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SEB_Swath_Width}
\alias{SEB_Swath_Width}
\title{A function to calculate the swath area viewed by a stereo camera}
\usage{
SEB_Swath_Width(horizontal_AOV, range_distance)
}
\arguments{
\item{range_distance}{the range at which to calculate the horizontal distance viewed}

\item{horizonal_AOV}{a horizontal viewing angle calculated from SEB_Viewing_Angles function or other source}
}
\description{
This function uses a horizonal viewing angle from SEB_Viewing_Angles to estimate the horizontal area
viewed by a stareo camera at a set range. Typically this is used to compute the area swepte along a
survey transect at a median range of objects that were observed. This method is based on Rooper et al
2016.
}
\examples{
swath<-SEB_Swath_Width(angles$horizontal_AOV, median(target_data$range,na.rm=TRUE))
}
\keyword{SEBASTES,}
\keyword{area}
\keyword{camera,}
\keyword{concatenation,}
\keyword{data}
\keyword{stereo}
\keyword{swept}
\keyword{wrangling}
