% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SEB_Viewing_Angles}
\alias{SEB_Viewing_Angles}
\title{A function to calculate the angle viewed by a stereo camera}
\usage{
SEB_Viewing_Angles(calibration_file, image_width, image_height, plot.it = TRUE)
}
\arguments{
\item{calibration_file}{a calibration file for a stereo camera from matlab (*.mat)}

\item{image_width}{the image width from the camera in pixels}

\item{image_height}{the image height from the camera in pixels}

\item{plot.it}{do you want to generate a plot of the viewing angle?}
}
\description{
This function uses a stereo camera calibration *.mat file to estimate the horizontal angle of view
for a camera pair. It is a translation from Matlab to R of functions from Kresimir Williams to
calculate horizontal AOV and uses a translated function from the Bouget et al. 2008 to do the  Rodrigues
transformation of the calibration.
}
\examples{
angles<-SEB_Viewing_Angles('dropcam_unit2_calibration_switched.mat', 2048,1536)
}
\keyword{SEBASTES,}
\keyword{area}
\keyword{camera,}
\keyword{concatenation,}
\keyword{data}
\keyword{stereo}
\keyword{swept}
\keyword{wrangling}
