% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{six_digit_framename_fix}
\alias{six_digit_framename_fix}
\title{A function to fix bad naming in profile for Sebastes deployment (happened in 2022 GOA and Seamount surveys)}
\usage{
six_digit_framename_fix(
  startDir,
  camera = c("Blackfly BFLY-PGE-50S5C_16396245", "Blackfly BFLY-PGE-50S5C_16304398")
)
}
\arguments{
\item{startDir}{Directory with the bad deployment data.}

\item{camera}{the name of one of the cameras in the stereo pari (default is DropCam1 'Blackfly BFLY-PGE-50S5C_16396245'}
}
\description{
This function fixes 5 digit numbering to 6 digit numbering and appendds new number to existing data
}
\examples{
six_digit_framename_fix("C:/Users/rooperc/Desktop/SeamountSebastes")
}
\keyword{SEBASTES,}
\keyword{camera,}
\keyword{concatenation,}
\keyword{data}
\keyword{stereo}
\keyword{wrangling}
